refactor: create a types file
