+++ src/components/Home/types.d.ts
@@ -1,7 +1,5 @@
 import type { DetailedHTMLProps, ImgHTMLAttributes } from 'react'
 import type { ButtonProps } from '@mui/material'
-import type { TypeBaseBlockSkeleton } from '@/contentful/types'
-import type { Entry } from 'contentful'
 
 type BaseBlock = {
   title: string | JSX.Element
@@ -24,5 +22,3 @@ type Button = {
   variant: 'button' | 'link'
   color?: ButtonProps['color']
 }
-
-type BaseBlockEntry = Entry<TypeBaseBlockSkeleton, undefined, string>
