+++ src/components/Token/Hero/index.tsx
@@ -5,7 +5,7 @@ import OrbitRingsList from '@/components/Token/Hero/token-hero-rings.json'
 import OrbitPath from '@/components/Token/OrbitPath'
 import Floater from '@/components/Token/Floaters'
 import PulsingLogo from '@/components/Token/PulsingLogo'
-import { type BaseBlockEntry } from '@/components/Home/types'
+import { type BaseBlockEntry } from '@/config/types'
 import RichText from '@/components/common/RichText'
 import ButtonsWrapper from '@/components/Token/ButtonsWrapper'
 import { isAsset, isEntryTypeButton } from '@/lib/typeGuards'
