+++ src/components/Token/ParallaxText/index.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from 'react'
 import { Container, Grid } from '@mui/material'
 import layoutCss from '@/components/common/styles.module.css'
 import css from './styles.module.css'
-import type { BaseBlockEntry } from '@/components/Home/types'
+import { type BaseBlockEntry } from '@/config/types'
 import RichText from '@/components/common/RichText'
 import clsx from 'clsx'
 import ButtonsWrapper from '@/components/Token/ButtonsWrapper'
