+++ src/components/Token/SafetyNumbers/index.tsx
@@ -1,5 +1,5 @@
 import RichText from '@/components/common/RichText'
-import { type BaseBlockEntry } from '@/components/Home/types'
+import { type BaseBlockEntry } from '@/config/types'
 import { Box, Container, Grid } from '@mui/material'
 import layoutCss from '@/components/common/styles.module.css'
 import Marquee from '@/components/common/Marquee'
