+++ src/components/Token/Tokenomics/index.tsx
@@ -1,7 +1,7 @@
 import { Box, Container, Grid } from '@mui/material'
 import layoutCss from '@/components/common/styles.module.css'
 import RichText from '@/components/common/RichText'
-import { type BaseBlockEntry } from '@/components/Home/types'
+import { type BaseBlockEntry } from '@/config/types'
 import css from './styles.module.css'
 import LinkButton from '@/components/common/LinkButton'
 import { isAsset, isEntryTypeBaseBlock, isEntryTypeExternalURL } from '@/lib/typeGuards'
