+++ src/components/common/LegalDisclaimer/index.tsx
@@ -1,6 +1,6 @@
 import { Container, Typography } from '@mui/material'
 import layoutCss from '@/components/common/styles.module.css'
-import { type BaseBlockEntry } from '@/components/Home/types'
+import { type BaseBlockEntry } from '@/config/types'
 import RichText from '@/components/common/RichText'
 
 const LegalDisclaimer = (props: BaseBlockEntry) => {
