+++ src/config/types.ts
@@ -1,0 +1,4 @@
+import type { TypeBaseBlockSkeleton } from '@/contentful/types'
+import type { Entry } from 'contentful'
+
+export type BaseBlockEntry = Entry<TypeBaseBlockSkeleton, undefined, string>
